LayoutTests/imported/w3c/web-platform-tests/css/cssom/font-shorthand-serialization-expected.txt


FAIL The font shorthand should be serialized just like any other shorthand. assert_equals: expected "font: 10px / 1 Ahem;" but got "font-style: initial; font-variant-caps: initial; font-weight: initial; font-stretch: initial; font-size: 10px; line-height: 1; font-family: Ahem; font-size-adjust: initial; font-kerning: initial; font-variant-alternates: initial; font-variant-ligatures: initial; font-variant-numeric: initial; font-variant-east-asian: initial; font-variant-position: initial; font-feature-settings: initial; font-optical-sizing: initial; font-variation-settings: initial; font-palette: initial;"
